fix: singlecmd
