+++ src/executecmd/run_redirection.c
@@ -6,7 +6,7 @@
 /*   By: sanghupa <sanghupa@student.42berlin.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/13 12:12:54 by minakim           #+#    #+#             */
-/*   Updated: 2023/11/17 14:11:37 by minakim          ###   ########.fr       */
+/*   Updated: 2023/11/23 14:34:35 by minakim          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
